# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERP API",
        "description": "List of all api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/department/store": {
            "post": {
                "tags": [
                    "Department"
                ],
                "summary": "Store department",
                "operationId": "debd957293f4df5d9ba8bc9814e62f72",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "IT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully department saved."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/department/show": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "List departments",
                "operationId": "8fa0ab1796e1905f47ffcd70de4e4308",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "numeric",
                                                        "example": "24"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "IT"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/department/edit/{id}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Edit department detail",
                "operationId": "f014365753fec809fc57f333022073b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "25"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get department detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "numeric",
                                                    "example": "25"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Finance"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/department/update": {
            "post": {
                "tags": [
                    "Department"
                ],
                "summary": "Update department",
                "operationId": "140a47513fc21ae621a8c8ac5f05cfc2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "numeric",
                                        "example": "25"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "IT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully department updated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/department/delete/{id}": {
            "delete": {
                "tags": [
                    "Department"
                ],
                "summary": "Delete department",
                "operationId": "0b76dc422cd66b0ff57d295ccdef0cfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the Department",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "25"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully department deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}